# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "BookmarkCreate": {
        "properties": {
          "entity_id": {
            "pattern": "^[a-z0-9][a-z0-9-]{0,159}$",
            "title": "Entity Id",
            "type": "string"
          }
        },
        "required": [
          "entity_id"
        ],
        "title": "BookmarkCreate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProfilePatch": {
        "properties": {
          "experience_level": {
            "anyOf": [
              {
                "pattern": "^(novice|beginner|intermediate|advanced)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Level"
          },
          "learning_pace": {
            "anyOf": [
              {
                "pattern": "^(light|steady|intensive)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learning Pace"
          },
          "target_difficulty": {
            "anyOf": [
              {
                "pattern": "^(beginner|intermediate|advanced)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Difficulty"
          }
        },
        "title": "ProfilePatch",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Versioned, local-first API over deterministic generated knowledge contracts, strictly bounded graph intelligence, and proposal-only community collaboration. Community routes never mutate canonical knowledge; Git and maintainer review remain the publication boundary. Lab routes accept only predefined safe local-fixture actions; arbitrary commands and target scanning are not supported.",
    "title": "Universal Hacking Tools Knowledge API",
    "version": "10.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me": {
      "delete": {
        "operationId": "delete_account_api_v1_me_delete",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "uht_csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete the authenticated account and all private application state",
        "tags": [
          "authenticated",
          "private"
        ]
      },
      "get": {
        "operationId": "me_api_v1_me_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read the authenticated user profile",
        "tags": [
          "authenticated",
          "private"
        ]
      },
      "patch": {
        "operationId": "patch_me_api_v1_me_patch",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "uht_csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update safe authenticated learner preferences",
        "tags": [
          "authenticated",
          "private"
        ]
      }
    },
    "/api/v1/me/achievements": {
      "get": {
        "operationId": "get_achievements_api_v1_me_achievements_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read deterministic private achievements",
        "tags": [
          "authenticated",
          "private"
        ]
      }
    },
    "/api/v1/me/bookmarks": {
      "get": {
        "operationId": "get_bookmarks_api_v1_me_bookmarks_get",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List private bookmarks",
        "tags": [
          "authenticated",
          "private"
        ]
      },
      "post": {
        "operationId": "create_bookmark_api_v1_me_bookmarks_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "uht_csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a private bookmark for existing knowledge",
        "tags": [
          "authenticated",
          "private"
        ]
      }
    },
    "/api/v1/me/bookmarks/{entity_id}": {
      "delete": {
        "operationId": "delete_bookmark_api_v1_me_bookmarks__entity_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "uht_csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete a private bookmark",
        "tags": [
          "authenticated",
          "private"
        ]
      }
    },
    "/api/v1/me/change-password": {
      "post": {
        "operationId": "change_password_api_v1_me_change_password_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "uht_csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change password and rotate authenticated session",
        "tags": [
          "authenticated",
          "private"
        ]
      }
    },
...
    "/api/v1/me/dashboard": {
      "get": {
        "operationId": "dashboard_api_v1_me_dashboard_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read a bounded private learning dashboard summary",
        "tags": [
          "authenticated",
          "private"
        ]
      }
    },
...
    "/api/v1/me/goals/{goal_id}": {
      "delete": {
        "operationId": "remove_goal_api_v1_me_goals__goal_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "goal_id",
            "required": true,
            "schema": {
              "title": "Goal Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "uht_csrf",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove one private learning goal",
        "tags": [
          "authenticated",
          "private"
        ]
      }
    },
    "/api/v1/me/knowledge-gaps": {
      "get": {
        "operationId": "knowledge_gaps_api_v1_me_knowledge_gaps_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "uht_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Uht Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read deterministic owner-scoped gaps for selected generated learning paths",
        "tags": [
          "authenticated",
          "private"
        ]
      }
    },
...
  }
}